requirements.txt
pylint
ldap3
python-dotenv
